Replace inappropriate instances of 'graph' with 'chart'. Specifically, renaming  graph_generator folder, take_graph_data.py to chart_generator, take_chart_data, changing functions, menu text, docstrings, comments etc.
