+++ README.md
@@ -1,4 +1,4 @@
-# dionysus - Avatar graph generator
+# dionysus - Avatar chart generator
 [![Build Status](https://travis-ci.org/toonarmycaptain/dionysus.svg?branch=master)](https://travis-ci.org/toonarmycaptain/dionysus) [![Build status](https://ci.appveyor.com/api/projects/status/yb33uwd13tkv7l79?svg=true)](https://ci.appveyor.com/project/toonarmycaptain/dionysus) [![Coverage Status](https://coveralls.io/repos/github/toonarmycaptain/dionysus/badge.svg)](https://coveralls.io/github/toonarmycaptain/dionysus) [![BCH compliance](https://bettercodehub.com/edge/badge/toonarmycaptain/dionysus?branch=master)](https://bettercodehub.com/) [![Codacy Badge](https://api.codacy.com/project/badge/Grade/d24e9508258849c2b40760fce3448c6b)](https://www.codacy.com/app/toonarmycaptain/dionysus?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=toonarmycaptain/dionysus&amp;utm_campaign=Badge_Grade) [![CodeFactor](https://www.codefactor.io/repository/github/toonarmycaptain/dionysus/badge/master)](https://www.codefactor.io/repository/github/toonarmycaptain/dionysus/overview/master) [![Updates](https://pyup.io/repos/github/toonarmycaptain/dionysus/shield.svg)](https://pyup.io/repos/github/toonarmycaptain/dionysus/) [![Known Vulnerabilities](https://snyk.io/test/github/toonarmycaptain/dionysus/badge.svg?targetFile=requirements.txt)](https://snyk.io/test/github/toonarmycaptain/dionysus?targetFile=requirements.txt) [![Python 3](https://pyup.io/repos/github/toonarmycaptain/dionysus/python-3-shield.svg)](https://pyup.io/repos/github/toonarmycaptain/dionysus/) 
 
 **dionysus** is an open source CLI app primarily aimed at teachers that charts student results for display using avatars, nicknames, or student names.
