+++ dionysus_app/chart_generator/create_chart.py
@@ -1,18 +1,18 @@
 """
 Script for composing data set to pass to graph_image_generator.
 
-Prototype will be a bar graph eg columns for each 10pt range 0-100, each avatar stacked on top of each other.
+Prototype will be a bar chart eg columns for each 10pt range 0-100, each avatar stacked on top of each other.
 
-Immediate enhancement from there will be variable ranges for the graph, columns eg 0-15 for a quiz rather than
-a percentage, or column widths of 5pts rather than 10. Other potential concern is graph being too high, so some
+Immediate enhancement from there will be variable ranges for the chart, columns eg 0-15 for a quiz rather than
+a percentage, or column widths of 5pts rather than 10. Other potential concern is chart being too high, so some
 sort of overlap without obscuring the avatars, or two columns of avatars in a point column.
 """
 from copy import deepcopy
 
 from dionysus_app.class_functions import select_classlist
 from dionysus_app.data_folder import CHART_DATA_FILE_TYPE, DataFolder
-from dionysus_app.graph_generator.generate_image import generate_chart_image
-from dionysus_app.graph_generator.take_graph_data import take_chart_name, take_score_data
+from dionysus_app.chart_generator.generate_image import generate_chart_image
+from dionysus_app.chart_generator.take_chart_data import take_chart_name, take_score_data
 from dionysus_app.file_functions import convert_to_json
 from dionysus_app.UI_functions import clean_for_filename
 
@@ -20,17 +20,17 @@ from dionysus_app.UI_functions import clean_for_filename
 CLASSLIST_DATA_PATH = DataFolder.generate_rel_path(DataFolder.CLASS_DATA.value)
 
 
-def new_graph():
+def new_chart():
     """
     TODO: write docstring. In particular form for chart_data_dict.
     :return:
     """
     # Select class or redirect to create new class.
-    # Take graph_name
-    # Take new data set. Store in class_data/{class_name}/graph_data/graph_name.cdf (ChartDataFile)
-    # Pass data to graph image creation scripts
+    # Take chart_name
+    # Take new data set. Store in class_data/{class_name}/chart_data/chart_name.cdf (ChartDataFile)
+    # Pass data to chart image creation scripts
     #   - potential options in those scripts (or here) to include:
-    #       - graph/image title options eg name different from displayed title
+    #       - chart/image title options eg name different from displayed title
     #       - axis labels, scale/axis tick markings
 
     chart_name, class_name, student_scores, chart_params = assemble_chart_data()
@@ -86,7 +86,7 @@ def write_chart_data_to_file(chart_data_dict: dict):
     file_chart_data_dict = deepcopy(chart_data_dict)
     chart_filename = clean_for_filename(file_chart_data_dict['chart_name'])
     chart_data_file = chart_filename + CHART_DATA_FILE_TYPE
-    chart_data_filepath = CLASSLIST_DATA_PATH.joinpath(file_chart_data_dict['class_name'], 'graph_data', chart_data_file)
+    chart_data_filepath = CLASSLIST_DATA_PATH.joinpath(file_chart_data_dict['class_name'], 'chart_data', chart_data_file)
 
     json_safe_chart_data_dict = sanitise_avatar_path_objects(file_chart_data_dict)
     json_chart_data = convert_to_json(json_safe_chart_data_dict)
