+++ dionysus_app/chart_generator/default_avatar.png
