+++ dionysus_app/main_menu.py
@@ -5,11 +5,11 @@ Application main menu.
 import sys
 
 from dionysus_app.class_functions import create_classlist
-from dionysus_app.graph_generator.create_graph import new_graph
+from dionysus_app.chart_generator.create_chart import new_chart
 
 
 def welcome_blurb():
-    print("Welcome to Dionysus - student avatar graph generator\n")
+    print("Welcome to Dionysus - student avatar chart generator\n")
 
 
 def main_menu_options():
@@ -17,7 +17,7 @@ def main_menu_options():
     print("Please select an option by entering the corresponding number, and press return:\n"
           "     1. Create a classlist\n"
           "     2. Edit a classlist\n"
-          "     3. Create a new graph\n"
+          "     3. Create a new chart\n"
           "     Enter Q to quit.\n")
 
 
@@ -33,7 +33,7 @@ def take_main_menu_input():
     possible_options = {
         '1': create_classlist,
         '2': 'edit_classlist',
-        '3': new_graph,
+        '3': new_chart,
         'q': quit_app,
         'Q': quit_app,
         }
