+++ test_suite/test_data_folder.py
@@ -11,8 +11,8 @@ class TestDataFolder(unittest.TestCase):
             r'/dionysus_app/app_data',
             r'/dionysus_app/app_data/class_data',
             r'/dionysus_app/app_data/image_data',
-            r'/dionysus_app/graph_generator',
-            r'/dionysus_app/graph_generator/default_avatar.png',
+            r'/dionysus_app/chart_generator',
+            r'/dionysus_app/chart_generator/default_avatar.png',
         ]
 
     def test_generate_data_path_defaults(self):
